delivery/requirements.txt
aio-pika==9.5.4
billiard==4.2.1
black==24.10.0
channels-rabbitmq==4.0.1
channels_redis==4.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework_simplejwt==5.4.0
google-auth==2.37.0
kombu==5.4.2
pika==1.3.2
pillow==11.1.0
pip==24.3.1
psycopg2-binary==2.9.10
pytest-django==4.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
Werkzeug==3.1.3
